move all files in a dedicated folder and use package feature
